+++ packages/plugin-packages/mashroom-monitoring-pm2-exporter/type-definitions/index.d.ts
@@ -1,0 +1,12 @@
+
+import type {MetricConfig} from '@pm2/io/build/main/features/metrics';
+
+export interface PM2MetricExporter {
+    start(): void;
+    stop(): void;
+}
+
+export type Config = {
+    pmxMetrics: MetricConfig;
+    mashroomMetrics: Array<string>;
+}